Extend signal handling to parsing for clean exit
