+++ src/Server.cpp
@@ -1,15 +1,13 @@
 #include "Server.hpp"
 #include "Queue.hpp"
 
-volatile sig_atomic_t g_ServerShoudClose = false;
+extern sig_atomic_t g_ShouldStop;
 
 static int	start_servers(const std::vector<Configuration>,Endpoint*,int,int*);
 static Endpoint	*connectNewClient(Endpoint *, const Endpoint *, int, int *);
 static void	initEndpoint(int, std::string, std::string, Endpoint *);
 static bool endpointAlreadyBound(Endpoint *, int, std::string, std::string);
 static bool	isTimedOut(Endpoint *, int);
-static void sigcleanup(int);
-static void handlesignals(void(*)(int));
 
 int	run(const std::vector<Configuration> config)
 {
@@ -19,8 +17,6 @@ int	run(const std::vector<Configuration> config)
     std::cerr << "Error: webserv: Too many virtual hosts\n";
     return 1;
   }
-
-	handlesignals(sigcleanup);
 
 	int	error = 1;
 
@@ -52,8 +48,8 @@ int	run(const std::vector<Configuration> config)
 	queue_event events[QUEUE_MAX_EVENTS];
 	bzero(events, sizeof(events));
 
-	while (!g_ServerShoudClose) {
-		assert(g_ServerShoudClose == false);
+	while (!g_ShouldStop) {
+		assert(g_ShouldStop == false);
 
 		for (Endpoint *conn = endpoints; conn < endpoints + max_client_id; conn++) {
 			if (isLiveClient(conn) && isTimedOut(conn, qfd)) {
@@ -88,7 +84,8 @@ int	run(const std::vector<Configuration> config)
 				 }
 					break;
 
-        case None: break;
+        case None: assert(false); /* Unreachable */
+          break;
 			}
 		}
 	}
@@ -252,26 +249,6 @@ static void	initEndpoint(int sockfd, std::string host, std::string port,
 	assert(strlen(endpoint->port) >= 1);
 	assert(strlen(endpoint->IP) >= 7);
 	assert(strlen(endpoint->IP) <= 16);
-}
-
-static void sigcleanup(int sig)
-{
-	(void)sig;
-	g_ServerShoudClose = true;
-}
-
-static void handlesignals(void(*hdl)(int))
-{
-	struct sigaction sa;
-
-	signal(SIGPIPE, SIG_IGN);
-	memset(&sa, 0, sizeof(sa));
-	sa.sa_handler = hdl;
-	sigemptyset(&sa.sa_mask);
-	sigaction(SIGTERM, &sa, nullptr);
-	sigaction(SIGHUP, &sa, nullptr);
-	sigaction(SIGINT, &sa, nullptr);
-	sigaction(SIGQUIT, &sa, nullptr);
 }
 
 int		watch(int qfd, Endpoint *conn, enum queue_event_type t)
